+++ packages/pangea-sdk/scripts/test.sh
@@ -10,6 +10,11 @@ pnpm dlx start-server-and-test --expect 404 \
   "poetry run pytest tests/integration2/test_ai_guard.py"
 
 pnpm dlx start-server-and-test --expect 404 \
+  "pnpm dlx @stoplight/prism-cli mock -d --json-schema-faker-fillProperties=false tests/testdata/specs/audit.openapi.json" \
+  4010 \
+  "poetry run pytest tests/integration2/test_audit.py"
+
+pnpm dlx start-server-and-test --expect 404 \
   "pnpm dlx @stoplight/prism-cli mock -d --json-schema-faker-fillProperties=false tests/testdata/specs/authn.openapi.json" \
   4010 \
   "poetry run pytest tests/integration2/test_authn.py"